# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Systock API",
        "description": "Systock Challenge API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Listar usuários",
                "operationId": "e9d4d8b419dbef04b5748fbf9ccf3881",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários retornada com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Criar novo usuário",
                "operationId": "4e9dbcbdaaaf9faab196a82d5cbaea23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cpf",
                                    "email",
                                    "senha",
                                    "tipo_usuario"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "minLength": 6
                                    },
                                    "tipo_usuario": {
                                        "description": "1: Admin, 2: Cliente",
                                        "type": "integer",
                                        "enum": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Buscar usuário por ID",
                "operationId": "e9ac05ddc4ea46d8848b012d7d92e3ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualizar usuário existente",
                "operationId": "27ac608b6dd8b8e7c7c45b10c5210b3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "tipo_usuario": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Remover usuário",
                "operationId": "ea700764859a665afd38bb39c2842f1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário removido com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}